+++ revanced-patcher/src/test/kotlin/app/revanced/patcher/patch/usage/ExampleResourcePatch.kt
@@ -1,0 +1,22 @@
+package app.revanced.patcher.patch.usage
+
+import app.revanced.patcher.data.ResourceContext
+import app.revanced.patcher.patch.ResourcePatch
+import org.w3c.dom.Element
+
+
+class ExampleResourcePatch : ResourcePatch() {
+    override fun execute(context: ResourceContext) {
+        context.xmlEditor["AndroidManifest.xml"].use { editor ->
+            val element = editor // regular DomFileEditor
+                .file
+                .getElementsByTagName("application")
+                .item(0) as Element
+            element
+                .setAttribute(
+                    "exampleAttribute",
+                    "exampleValue"
+                )
+        }
+    }
+}