+++ src/main/kotlin/app/revanced/patcher/PatchesConsumer.kt
@@ -1,8 +1,0 @@
-package app.revanced.patcher
-
-import app.revanced.patcher.patch.PatchClass
-
-@FunctionalInterface
-interface PatchesConsumer {
-    fun acceptPatches(patches: List<PatchClass>)
-}